Moved Validate to existing class.
